Add a millisecond timer
Accurate to 90% with WS2812B's interrupt disabling and a 15ms delay, or
accurate to 95% without disabling interrupts. I assume any drift here is
mainly because of the inaccuracy of the microcontroller's internal
oscillator, but 90% is fine: the missing time can be spread out
over the intervening time period for good enough animations.